refactor(agent): split login helpers and preserve owners
