+++ npa/tests/test_config.py
@@ -709,9 +709,8 @@ def test_remove_workbench_config_updates_defaults(isolated_config: Path) -> None
 
     config.remove_workbench_config("proj-a", "wb-b")
     data = yaml.safe_load(isolated_config.read_text())
-    assert "workbenches" not in data["projects"]["proj-a"]
-    assert data["projects"]["proj-a"]["project_id"] == "project-1"
-    assert data["default_project"] == "proj-a"
+    assert "proj-a" not in data["projects"]
+    assert data["default_project"] == "proj-b"
 
 
 def test_remove_last_workbench_preserves_agent_ownership_record(
